+++ src/main/java/kr/fiveminutesmarket/common/config/SecurityConfig.java
@@ -1,5 +1,7 @@
 package kr.fiveminutesmarket.common.config;
 
+import kr.fiveminutesmarket.common.interceptor.AuthenticationInterceptor;
+import kr.fiveminutesmarket.common.interceptor.SessionExtensionInterceptor;
 import kr.fiveminutesmarket.common.security.JavaPasswordEncoder;
 import kr.fiveminutesmarket.common.security.SHA256PasswordEncoder;
 import kr.fiveminutesmarket.common.security.SecurityUserResolver;
@@ -7,6 +9,7 @@ import kr.fiveminutesmarket.common.utils.RedisAuthUtils;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.List;
@@ -25,8 +28,16 @@ public class SecurityConfig implements WebMvcConfigurer {
         return new SHA256PasswordEncoder();
     }
 
+    // @LoginUser param object 주입을 위한 resolver
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new SecurityUserResolver(redisAuthUtils));
+        resolvers.add(new SecurityUserResolver());
+    }
+
+    // Interceptor related to Session Authentication
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new SessionExtensionInterceptor(redisAuthUtils));
+        registry.addInterceptor(new AuthenticationInterceptor());
     }
 }
